DEV: Use pick-files-input in composer-editor and clean up
A while ago in 27b97e4f648da1587e97c5f661da145e12a27bea the
pick-files-input was added but never used. This commit uses it
for the composer-editor, and cleans it up to remove unnecessary
binding and validation logic which is all handled by uppy now.

This can then be used in other places in the app and also for plugins.
