+++ app/assets/javascripts/discourse/app/templates/components/composer-editor.hbs
@@ -22,10 +22,4 @@
   {{yield}}
 {{/d-editor}}
 
-{{#if allowUpload}}
-  {{#if acceptsAllFormats}}
-    <input type="file" id="file-uploader" multiple>
-  {{else}}
-    <input type="file" id="file-uploader" accept={{acceptedFormats}} multiple>
-  {{/if}}
-{{/if}}
+{{pick-files-input allowUpload=allowUpload fileInputId="file-uploader" allowMultiple=true}}
