--- app/assets/javascripts/discourse/app/templates/components/pick-files-button.hbs
@@ -1,6 +1,0 @@
-{{d-button action=(action "openSystemFilePicker") label=label icon=icon}}
-{{#if acceptAnyFile}}
-  <input type="file">
-{{else}}
-  <input type="file" accept={{acceptedFileTypesString}}>
-{{/if}}
