+++ app/assets/javascripts/discourse/app/templates/components/pick-files-input.hbs
@@ -1,0 +1,10 @@
+{{#if allowUpload}}
+  {{#if showButton}}
+    {{d-button action=(action "openSystemFilePicker") label=label icon=icon}}
+  {{/if}}
+  {{#if acceptsAllFormats}}
+    <input type="file" id={{fileInputId}} class={{fileInputClass}} multiple={{allowMultiple}}>
+  {{else}}
+    <input type="file" id={{fileInputId}} class={{fileInputClass}} accept={{acceptedFileTypesString}} multiple={{allowMultiple}}>
+  {{/if}}
+{{/if}}
